:wrench: 타입 파일 이동, 변경된 아이콘 적용
