--- src/components/PreviewCarousel/PreviewCarouselModal/PreviewCarouselModal.type.ts
@@ -1,7 +1,0 @@
-export interface IPreviewCarouselModal {
-  selectedIndex: number;
-  thumbnailListState: Array<string>;
-  handleLeftClick: () => void;
-  handleRightClick: () => void;
-  handleModalClose: () => void;
-}
